OCR-Python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
easyocr==1.7.0
python-docx==1.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
pytesseract==0.3.10
reportlab==4.0.4
